requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
browser-cookie3==0.19.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
exceptiongroup==1.2.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-text-splitters==0.3.2
langsmith==0.1.143
loguru==0.7.2
lz4==4.3.3
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.11
packaging==24.2
pinecone==5.4.0
pinecone-plugin-inference==2.0.1
pinecone-plugin-interface==0.0.7
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
protoc-gen-openapiv2==0.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gemini-api==2.4.12
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==9.0.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
